feat: add settlement amount
