+++ src/main/kotlin/com/pravo/pravo/domain/promise/service/PromiseFacade.kt
@@ -1,6 +1,9 @@
 package com.pravo.pravo.domain.promise.service
 
 import com.pravo.pravo.domain.member.service.MemberService
+import com.pravo.pravo.domain.point.service.PointLogService
+import com.pravo.pravo.domain.promise.dto.response.ParticipantResponseDto
+import com.pravo.pravo.domain.promise.dto.response.PromiseDetailResponseDto
 import com.pravo.pravo.domain.promise.dto.response.PromiseResponseDto
 import com.pravo.pravo.domain.promise.model.enums.ParticipantStatus
 import com.pravo.pravo.domain.promise.model.enums.PromiseStatus
@@ -13,6 +16,7 @@ class PromiseFacade(
     private val promiseService: PromiseService,
     private val promiseRoleService: PromiseRoleService,
     private val memberService: MemberService,
+    private val pointLogService: PointLogService,
 ) {
     @Transactional
     fun joinPromise(
@@ -38,5 +42,25 @@ class PromiseFacade(
 
         val promiseRole = promiseRoleService.getPromiseRole(memberId, promiseId)
         promiseRole.updateStatus(ParticipantStatus.READY)
+    }
+
+    fun getPromiseDetailByMember(
+        memberId: Long,
+        promiseId: Long,
+    ): PromiseDetailResponseDto {
+        val promise =
+            promiseService.getPromiseWithFetch(promiseId)
+        val participants =
+            promise.promiseRoles
+                .filter { it.status != ParticipantStatus.PENDING }
+                .map {
+                    ParticipantResponseDto.of(it, it.member)
+                }
+        val settlementAmount =
+            when (promise.status) {
+                PromiseStatus.COMPLETED -> pointLogService.calculateSettlementAmount(memberId, promiseId, promise.deposit)
+                else -> null
+            }
+        return PromiseDetailResponseDto.of(promise, participants, settlementAmount)
     }
 }
