+++ src/main/kotlin/com/pravo/pravo/domain/promise/service/PromiseService.kt
@@ -1,11 +1,8 @@
 package com.pravo.pravo.domain.promise.service
 
 import com.pravo.pravo.domain.promise.dto.request.PromiseCreateDto
-import com.pravo.pravo.domain.promise.dto.response.ParticipantResponseDto
-import com.pravo.pravo.domain.promise.dto.response.PromiseDetailResponseDto
 import com.pravo.pravo.domain.promise.dto.response.PromiseResponseDto
 import com.pravo.pravo.domain.promise.model.Promise
-import com.pravo.pravo.domain.promise.model.enums.ParticipantStatus
 import com.pravo.pravo.domain.promise.repository.PromiseRepository
 import com.pravo.pravo.global.error.ErrorCode
 import com.pravo.pravo.global.error.exception.NotFoundException
@@ -22,23 +19,6 @@ class PromiseService(
         endedAt: LocalDate?,
     ): List<PromiseResponseDto> = promiseRepository.getPromisesByMemberIdAndStartedAtAndEndedAt(memberId, startedAt, endedAt)
 
-    fun getPromiseDetailByMember(
-        memberId: Long,
-        promiseId: Long,
-    ): PromiseDetailResponseDto {
-        val promise =
-            promiseRepository.getPromiseById(promiseId).orElseThrow {
-                NotFoundException(ErrorCode.BAD_REQUEST, "약속을 찾을 수 없습니다")
-            }
-        val participants =
-            promise.promiseRoles
-                .filter { it.status != ParticipantStatus.PENDING }
-                .map {
-                    ParticipantResponseDto.of(it, it.member)
-                }
-        return PromiseDetailResponseDto.of(promise, participants)
-    }
-
     fun createPendingPromise(promiseCreateDto: PromiseCreateDto): Promise {
         val promise = Promise.pendingOf(promiseCreateDto)
         return promiseRepository.save(promise)
@@ -46,6 +26,11 @@ class PromiseService(
 
     fun getPromise(promiseId: Long): Promise {
         return promiseRepository.findById(promiseId)
+            .orElseThrow { NotFoundException(ErrorCode.BAD_REQUEST, "약속을 찾을 수 없습니다") }
+    }
+
+    fun getPromiseWithFetch(promiseId: Long): Promise {
+        return promiseRepository.getPromiseById(promiseId)
             .orElseThrow { NotFoundException(ErrorCode.BAD_REQUEST, "약속을 찾을 수 없습니다") }
     }
 }
